requirements.txt
absl-py==2.1.0
accelerate==1.2.1
accelerated-scan==0.2.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==1.0.0
alembic==1.14.0
altair==5.5.0
aniso8601==10.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asciitree==0.3.3
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrdict==2.0.1
attrs==24.3.0
audioread==3.0.1
av==14.0.1
babel==2.16.0
beautifulsoup4==4.12.3
better-profanity==0.7.0
bitsandbytes==0.45.0
black==24.10.0
blinker==1.9.0
boto3==1.36.2
botocore==1.36.2
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.5.1
calflops==0.3.2
cdifflib==1.2.6
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clip==0.2.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cryptography==42.0.8
cupy-cuda12x==13.4.0
cycler==0.12.1
Cython==3.0.11
cytoolz==1.0.1
DataProperty==1.1.0
datasets==3.2.0
decorator==5.1.1
decord==0.6.0
denoising-diffusion-pytorch==2.1.1
diffusers==0.32.2
dill==0.3.8
Distance==0.1.3
Django==4.2.15
django-auditlog==3.0.0
django-filter==23.5
django-js-asset==3.0.1
django-mptt==0.14.0
djangorestframework==3.15.1
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
drf-excel==2.4.0
drf-flex-fields==1.0.2
editdistance==0.8.1
einops==0.7.0
einops-exts==0.0.4
einx==0.1.3
ema-pytorch==0.7.7
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0.post1
-e git+https://github.com/Dao-AILab/fast-hadamard-transform.git@d1a56eee9d502e67faacf61b7b947180d66b32a0#egg=fast_hadamard_transform
fastapi==0.115.6
fasteners==0.19
fastrlock==0.8.3
fasttext==0.9.3
fiddle==0.3.0
filelock==3.16.1
Flask==3.1.0
Flask-RESTful==0.3.10
flatbuffers==25.2.10
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
future==1.0.0
g2p-en==2.1.0
gast==0.6.0
gdown==5.2.0
geoopt==0.5.0
gevent==24.11.1
geventhttpclient==2.0.2
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.25.2
hydra-core==1.3.2
idna==3.10
ijson==3.3.0
imageio==2.36.1
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.0.0
intervaltree==3.1.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.31.0
isort==5.13.2
itsdangerous==2.2.0
Janome==0.5.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiwer==3.0.5
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaldi-python-io==1.2.2
kaldiio==2.18.0
keras==3.9.0
kittycad==0.7.6
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
latexcodec==3.0.0
lazy_loader==0.4
ldap3==2.9.1
Levenshtein==0.26.1
lhotse==1.29.0
libclang==18.1.1
libcst==1.6.0
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.9
lilcom==1.8.0
llvmlite==0.43.0
lm_eval==0.4.7
lmdb==1.6.2
loguru==0.7.3
lpips==0.1.4
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdurl==0.1.2
mediapy==1.1.6
MedPy==0.5.2
Megatron==0.5.1
megatron-energon==4.0.0
megatron_core==0.9.0
ml-dtypes==0.4.1
monai==1.4.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.23.0
NEMO==6.0.3
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@1218a040c6a95451dc9567068e5ae5de8ff5633c
nemo_text_processing==1.1.0
nerfacc==0.5.3
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numcodecs==0.13.1
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
nvidia-pytriton==0.2.5
nvitop==1.4.2
omegaconf==2.3.0
onnx==1.17.0
open-clip-torch==2.24.0
OpenCC==1.1.9
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.3
peft==0.14.0
pesq==0.0.4
pexpect==4.9.0
pillow==10.3.0
plac==1.4.3
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
portalocker==3.1.1
prettytable==3.12.0
progress==1.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==19.0.0
pyasn1==0.6.1
pybind11==2.13.6
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.22
pydantic==2.10.5
pydantic-extra-types==2.10.3
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
PyMCubes==0.1.6
pymodbus==3.3.2
pymongo==4.11.3
pynini==2.1.6.post1
pynvml==12.0.0
pyparsing==3.2.1
pypinyin==0.53.0
pypinyin-dict==0.9.0
PySocks==1.7.1
pystoi==0.4.1
pytablewriter==1.2.1
pytest==8.3.4
pytest-mock==3.14.0
pytest-runner==6.0.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-rapidjson==1.20
pytorch-fid==0.3.0
pytorch-lightning==1.4.2
pytorch-msssim==1.0.0
pytorch_retinaface @ git+https://github.com/NVlabs/Pytorch_Retinaface.git@b843f45942d7693113b3073fab8ee741d4e655d9
pytz==2024.1
PyYAML==6.0.2
pyzmq==23.2.1
RapidFuzz==3.11.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rotary-embedding-torch==0.8.6
rouge_score==0.1.2
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==0.4.2
s3transfer==0.11.1
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
sh==1.14.3
shellingham==1.5.4
SimpleITK==2.4.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.13.0
soupsieve==2.6
sox==1.5.0
soxr==0.5.0.post1
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.3
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.37
sqlitedict==2.1.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.3
tabledata==1.3.4
tabulate==0.9.0
taming-transformers==0.0.1
tcolorpy==0.1.7
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.1
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.71
termcolor==2.5.0
text-unidecode==1.3
textdistance==4.6.3
texterrors==0.5.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.7.0
timm==1.0.13
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.4.1
torchaudio==2.4.1
torchdiffeq==0.2.5
torchmetrics==1.6.1
torchsde==0.2.6
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
trampoline==0.1.2
transformer_engine==1.13.0
transformer_engine_cu12==1.13.0
transformer_engine_torch==1.13.0
transformers==4.45.0
trimesh==4.5.3
triton==3.0.0
tritonclient==2.33.0
typeguard==4.4.1
typepy==1.3.4
typer==0.15.1
typing-inspect==0.6.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
visdom==0.2.4
wandb==0.19.4
watchdog==6.0.0
wcwidth==0.2.13
webdataset==0.2.100
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
word2number==1.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zarr==2.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
